refactor: Split into multiple files
