+++ src/lib/pai.ts
@@ -1,0 +1,39 @@
+export const PAI_MAP: { [key: string]: number } = {
+  m0: 51, // 赤五萬
+  m1: 11,
+  m2: 12,
+  m3: 13,
+  m4: 14,
+  m5: 15,
+  m6: 16,
+  m7: 17,
+  m8: 18,
+  m9: 19,
+  p0: 52, // 赤五筒
+  p1: 21,
+  p2: 22,
+  p3: 23,
+  p4: 24,
+  p5: 25,
+  p6: 26,
+  p7: 27,
+  p8: 28,
+  p9: 29,
+  s0: 53, // 赤五索
+  s1: 31,
+  s2: 32,
+  s3: 33,
+  s4: 34,
+  s5: 35,
+  s6: 36,
+  s7: 37,
+  s8: 38,
+  s9: 39,
+  z1: 41, // 東
+  z2: 42, // 南
+  z3: 43, // 西
+  z4: 44, // 北
+  z5: 45, // 白
+  z6: 46, // 發
+  z7: 47, // 中
+} as const;
