+++ src/lib/utils.ts
@@ -1,0 +1,4 @@
+export const assertNever = (arg: never): never => {
+  const value = JSON.stringify(arg);
+  throw new Error(`Expected code to be unreachable, but got: ${value}`);
+};
